% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteDistMain.R
\name{AggByAgeSexKpOuMechSt}
\alias{AggByAgeSexKpOuMechSt}
\title{AggBySexKpOuMechSt(data)}
\usage{
AggByAgeSexKpOuMechSt(data)
}
\arguments{
\item{data}{Input data which is read and processed to calculate the aggregate value}
}
\value{
Aggregated data for the input with summarized Value
}
\description{
Calculates the aggregated data and returns it depending on the options present
}
