% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{AllSitesList}
\alias{AllSitesList}
\title{AllSitesList(assignments, org_data)}
\usage{
AllSitesList(assignments, org_data)
}
\arguments{
\item{assignments}{dataframe - contains data of each country, their planning, community, facility and country level: /api/sqlViews/kEtZ2bSQCu2/data.json}

\item{org_data}{dataframe - List of all sites and their data with level names and UIDs, as well as site levels: api/dataStore/dataSetAssignments/ous}
}
\value{
dataframe containing all sites and their site types, their psnu names, psnu uids and country/id
}
\description{
Takes the assignments table containing all the countries and their site levels for different types
as well as the orghierarchy data containing all the sites, to combine and give site type and psnu name, id for each site
}
