% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteDistMain.R
\name{BuildDimensionLists}
\alias{BuildDimensionLists}
\title{BuildDimensionLists(data_element_map_item, dim_item_sets, mechanisms, country_details)}
\usage{
BuildDimensionLists(data_element_map_item, dim_item_sets, mechanisms,
  country_details)
}
\arguments{
\item{data_element_map_item}{Each row of data_element_map being sliced and passed}

\item{dim_item_sets}{Dataframe containing all the dimension item sets}

\item{mechanisms}{All historic mechanisms for the country filtered by id}

\item{country_details}{Country levels for a country}
}
\value{
List of dimensions for the analytics call
}
\description{
# get list of dimensions (parameters) for analytics call by level {planning, community, facility}
analytics will be called seperatly for each level
}
