% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteDistMain.R
\name{CalculateSiteDensities}
\alias{CalculateSiteDensities}
\title{CalculateSiteDensities(data_element_map, country_details, mechanisms, dim_item_sets, base_url, cores)}
\usage{
CalculateSiteDensities(data_element_map, country_details, mechanisms,
  dim_item_sets, base_url, cores = 1)
}
\arguments{
\item{data_element_map}{Dataframe to be processed with CalculateSiteDensity}

\item{country_details}{Country levels for a country}

\item{mechanisms}{All historic mechanisms for the country filtered by id}

\item{dim_item_sets}{Dataframe containing all the dimension item sets}

\item{base_url}{Refers to instance of datim being accessed}

\item{cores}{Number of cores to use for parallel execution}
}
\value{
Data containing all site densities for the parameters passed
}
\description{
Wrapper function to functionalize and parallelize calls to CalculateSiteDensity
}
