% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteDistMain.R
\name{CalculateSiteDensity}
\alias{CalculateSiteDensity}
\title{CalculateSiteDensity(data_element_map_item, country_details, mechanisms, dim_item_sets, base_url)}
\usage{
CalculateSiteDensity(
  data_element_map_item,
  country_details,
  mechanisms,
  dim_item_sets,
  base_url
)
}
\arguments{
\item{data_element_map_item}{Each row of data_element_map being sliced and passed}

\item{country_details}{Country levels for a country}

\item{mechanisms}{All historic mechanisms for the country filtered by id}

\item{dim_item_sets}{Dataframe containing all the dimension item sets}

\item{base_url}{Refers to instance of datim being accessed}
}
\value{
Dataframe containing site density for the parameters passed
}
\description{
Calculates historic distribution for each target, DSD/TA, and site given psnu/IM
}
