% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{checkMissingIndicators}
\alias{checkMissingIndicators}
\title{checkMissingIndicators(model)}
\usage{
checkMissingIndicators(model)
}
\arguments{
\item{model}{a datapack model}
}
\value{
a data frame with the indicator code and a row count for exisitng data in model
}
\description{
A function that returns missing indicators in schema not in model
}
