% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{checkModelDisagg}
\alias{checkModelDisagg}
\title{checkModelDisagg(model, disagg)}
\usage{
checkModelDisagg(model, disagg)
}
\arguments{
\item{model}{a datapack model}

\item{disagg}{a disagg type}
}
\value{
a data frame with the indicator code and a msg for investigation
}
\description{
A function that checks the model against schema
}
