% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteDistMain.R
\name{CheckSiteToolData}
\alias{CheckSiteToolData}
\title{CheckSiteToolData(d, d_old)}
\usage{
CheckSiteToolData(d, d_old = NULL, issue_error = FALSE)
}
\arguments{
\item{d}{list - the output of DistributeToSites}

\item{d_old}{list - a prior output of DistributeToSites
if present check includes comparison
all_equal(d_new$data$site$distributed, d_old$data$site$distributed)}

\item{issue_error}{boolean - if true failing one or more tests 
will return an error}
}
\value{
list with various check outcomes
}
\description{
Performs some checks on the results of distributing data to sites
}
