% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{CompareNamesIds}
\alias{CompareNamesIds}
\title{CompareNamesIds(names, ids, type)}
\usage{
CompareNamesIds(names, ids, type)
}
\arguments{
\item{names}{string vector - names of spcific class of metadata - category option, indicator etc}

\item{ids}{string vector - ids of specific class of metadata - category option, indicator etc}

\item{type}{string - metadata endpoint - cataegoryOptions, indicators, etc}
}
\value{

}
\description{
Checks name list and paired id list (same length) and verifies they correspond to each other
}
