% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{createDissagReport}
\alias{createDissagReport}
\title{createDissagReport(model, cop_year)}
\usage{
createDissagReport(model, cop_year)
}
\arguments{
\item{model}{a datapack model}

\item{disagg}{a cop year}
}
\value{
a list with 3 reports of different granularity on disagg mismatches
}
\description{
creates a report showing disagg msimatches and other visual aids
}
