% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPackCommons.R
\docType{package}
\name{datapackcommons}
\alias{datapackcommons}
\title{\code{datapackcommons} package}
\description{
Parsing and import functions for the ICPI Data Pack
https://github.com/jennybc/googlesheets/blob/master/R/googlesheets.R
}
