% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{diffDataEntryForm}
\alias{diffDataEntryForm}
\title{diffDataEntryForm(uid_a, uid_b)}
\usage{
diffDataEntryForm(
  uid_a,
  uid_b,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{d2_session}{a d2_session}

\item{dataset_a}{a dataset uid}

\item{dataset_b}{a dataset uid}
}
\value{
a list of all three differences
}
\description{
A function that compares two data entry forms based on uid inputs
that access sqlviews
}
