% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{diffSnuximModels}
\alias{diffSnuximModels}
\title{diffSnuximModels(model_old, model_new, full_diff = TRUE)}
\usage{
diffSnuximModels(
  model_old,
  model_new,
  full_diff = TRUE,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{model_old}{list - snuxim model}

\item{model_new}{list - snuxim model}

\item{full_diff}{If TRUE compares model difference fully}
}
\value{
deltas representing difference in data of both models
}
\description{
A function that compares old and new snuxim models
}
