% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteDistMain.R
\name{DropSitesFromDensity}
\alias{DropSitesFromDensity}
\title{DropSitesFromDensity}
\usage{
DropSitesFromDensity(site_density, sites)
}
\arguments{
\item{site_density}{dataframe from density object}

\item{sites}{character vector - list of site uids to KEEP}
}
\value{
Returns original density with two additional columns: dropped_site_reduction and
psnuValueH_after_site_drop. Additionally recalculated the percent column if sites are 
dropped.
}
\description{
Drops sites from a density data frame if they are not part of a list in sites,
recalculates psnuValueH and percent (density) based on dropping the sites
}
