% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenDataPackModel.R
\name{flattenDataPackModel_21}
\alias{flattenDataPackModel_21}
\title{flattenDataPackModel_21}
\usage{
flattenDataPackModel_21(data, country_uids = NULL)
}
\arguments{
\item{data}{list object with COP 21 model data}

\item{country_uids}{string vector listing country uids at the top level of the list hierarchy.}
}
\value{
list object with one result set for each country named by country uid
}
\description{
Takes a COP 21 Data Pack Model data file and removes non-results content and flattens
the results hierarchy
}
