% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{Get19TMechanisms}
\alias{Get19TMechanisms}
\title{Get19TMechanisms(base_url)}
\usage{
Get19TMechanisms(base_url)
}
\arguments{
\item{base_url}{string - base address of instance (text before api/ in URL)}
}
\description{
gets data from SQL view vtNAsZcMZiU - mechanisms with 2018Oct data
}
