% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{GetAnalyticsByDimensions}
\alias{GetAnalyticsByDimensions}
\title{export
param technical_area type string - technical area data element group name
param numerator_or_denominator type string - numerator / denominator data element group name
param disagg_to_group_set type string - numerator / denominator data element group name
disagg_element_to_dimension
param disagg_element_to_dimension type tibble cols(
disaggregation_type_element = col_character(),
data_dimension = col_character()
)}
\usage{
GetAnalyticsByDimensions(technical_area, numerator_or_denominator,
  disaggregation_type, targets_or_results, organisation_unit_ids, period,
  additional_dimensions = NULL, support_types = c("TA", "DSD"))
}
\description{
export
param technical_area type string - technical area data element group name
param numerator_or_denominator type string - numerator / denominator data element group name
param disagg_to_group_set type string - numerator / denominator data element group name
disagg_element_to_dimension
param disagg_element_to_dimension type tibble cols(
disaggregation_type_element = col_character(),
data_dimension = col_character()
)
}
