% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{getDatasetUids}
\alias{getDatasetUids}
\title{getDatasetUids}
\usage{
getDatasetUids(fiscal_year, type)
}
\arguments{
\item{fiscal_year}{character - one of {"19", "20", "21"}}

\item{type}{character - one of {"targets", "results", "subnat_impatt"}}
}
\value{
returns a character vector of the related dataset uids
}
\description{
returns character verctor of dataset uids for a given FY {"19", "20", "21"}
and type {"targets", "results"}
}
