% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{GetDataValueSet}
\alias{GetDataValueSet}
\title{GetDataValueSet(org_unit,data_set,start_date,end_date,children)}
\usage{
GetDataValueSet(org_unit, data_set, start_date, end_date,
  children = "true")
}
\arguments{
\item{org_unit}{UID of the organisation unit of interest}

\item{data_set}{UID of the dataset of interest}

\item{start_date}{Start date in YYYY-MM-DD format}

\item{end_date}{End date in YYYY-MM-DD format}

\item{children}{Boolean flag to indicate whether data of the Orgunits c
children should be returned.}
}
\value{
Only errors out if the file is not readable, does not exist or the
user cannot login to the server
}
\description{
Gets a data value set for the provided paramaters
}
