% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{GetDataWithIndicator}
\alias{GetDataWithIndicator}
\title{GetDataWithIndicator(indicator, org_units, level,
period, additional_dimensions, additional_filters)}
\usage{
GetDataWithIndicator(base_url, indicator, org_units, level, periods,
  additional_dimensions = NULL, additional_filters = NULL)
}
\arguments{
\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{indicator}{string - uid of indicator}

\item{org_units}{list of strings - list of org unit uids, operate as boundry org units}

\item{level}{string - org hierarchy level, only one allowed}

\item{periods}{list of strings - periods specified as required by DHIS2 analytics endpoint, 
currently only one period supported}

\item{additional_dimensions}{2 column dataframe, first column containing dimension item uids
and second column the related dimension uid. Dimensions appear as columns in output with names
based on dimension name in DHIS2.}

\item{additional_filters}{2 column dataframe, first column containing dimension item uids
and second column the related dimension uid.filters do not appear explicitly in output.}
}
\value{
A list with $time = time the function was called, 
$api_call = api call used, and 
$results = the data returnd by the analytics call
}
\description{
Gets data from DHIS2 using a single indicator
}
