% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{LoadConfig}
\alias{LoadConfig}
\title{LoadConfig(config_path)}
\usage{
LoadConfig(config_path = NA)
}
\arguments{
\item{config_path}{Path to the DHIS2 credentials file}
}
\value{
Only errors out if the file is not readable, does not exist or the
user cannot login to the server
}
\description{
Loads a JSON configuration file to access a DHIS2 instance
}
