% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{MilitarySiteData}
\alias{MilitarySiteData}
\title{MilitarySiteData(flat_ouhierarchy)}
\usage{
MilitarySiteData(flat_ouhierarchy)
}
\arguments{
\item{flat_ouhierarchy}{- flat_ouhierarchy dataframe - List of all sites and their data 
with level names and UIDs, as well as site levels: /api/sqlViews/kEtZ2bSQCu2/data.json}
}
\value{
dataframe with all military sites and their PSNU names, IDs
}
\description{
Takes all site data and filters out the military sites, generating their PSNU name and ID as well
}
