% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{pivotSchemaCombos}
\alias{pivotSchemaCombos}
\title{pivotSchemaCombos(cop_year)}
\usage{
pivotSchemaCombos(schema = NULL, cop_year = NULL)
}
\arguments{
\item{schema}{optional param if you wish to enter a manual schema}

\item{cop_year}{the specific cop year for the schema}
}
\value{
a pivoted schema
}
\description{
grabs the pertinent schema and pivots data long
}
