% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{RenameDimensionColumns}
\alias{RenameDimensionColumns}
\title{RenameDimensionColumns(data, type)}
\usage{
RenameDimensionColumns(data, type)
}
\arguments{
\item{data}{the unique dim_cop_type that is passed in the MapDimToOptions method}

\item{type}{It will pre-pend the string in type to the columns names}
}
\value{
The dataframe with renamed column names for dimensions
}
\description{
Renames the original column names of datapackcommons::dim_items_sets, 
by prepending the string in the type parameter
}
