% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{RetryAPI}
\alias{RetryAPI}
\title{RetryAPI(api_url, content_type, max_attempts)}
\usage{
RetryAPI(api_url, content_type, max_attempts = 3, timeout = 300)
}
\arguments{
\item{api_url}{string - full url for web request}

\item{content_type}{string - expected type of content in reposne e.d 'application/json'}

\item{max_attempts}{integer - maximum number of retries for succesful request}

\item{timeout}{integer - maximum time to wait for API response}
}
\value{
full api response
}
\description{
Submits specified api request up to specified maximum times
stopping when expected content type is returned with 200 response
}
