% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{StackPrefixedCols}
\alias{StackPrefixedCols}
\title{StackPrefixedCols(data, prefixes)}
\usage{
StackPrefixedCols(data, prefixes)
}
\arguments{
\item{data}{dataframe - contains data to stack}

\item{prefixes}{string vector - list of prefixes to include in column selection and stacking}
}
\value{
tibble with superset of columns without prefixes in column names
}
\description{
Takes columns from data with specified prefixes and stacks them based on the unprefixed
portion of the name. Columns not containing one of the prefixes are excluded in returned data.
}
