% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteDistMain.R
\name{TransformAnalyticsOutput_SiteTool}
\alias{TransformAnalyticsOutput_SiteTool}
\title{TransformAnalyticsOutput_SiteTool()}
\usage{
TransformAnalyticsOutput_SiteTool(
  analytics_results,
  dim_item_sets,
  data_element_map_item,
  planning_level
)
}
\arguments{
\item{analytics_results}{- data frame - this is a results list item returned from
calling GetData_Analytics and corresponds to the data element map item}

\item{dim_item_sets}{Dataframe containing all the dimension item sets 
e.g. datapackcommons::dim_item_sets}

\item{data_element_map_item}{Single row of data_element_map being sliced and passed}

\item{planning_level}{the planning level to use when pulling psnuid from 
analytics_results$ou_hierarchy}
}
\description{
Transforms API results into a table ready for downstream processing.
Takes API analytics output (which is disaggrepgated by dimensions 
and dimension items) and maps to cateogry options needed in output. This
includes splitting or aggregating dimension items related to age, sex, kp, and other_disagge 
as specified in config files, dim_item_sets and data element map.

Additionally creates a psnuid column from the ou_hierarchy column and drops
the ou_hierarchy column.
}
