% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{translateDims}
\alias{translateDims}
\title{Translate Dimensions}
\usage{
translateDims(dimensions_df)
}
\arguments{
\item{dimensions_df}{a dimensions data frame.}
}
\value{
List of dimensions for the analytics call to datimutils::getAnalytics
}
\description{
takes a dimensions data frame and translates it into a list item for input to get_analytics
}
