% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{translateFils}
\alias{translateFils}
\title{Translate Filters}
\usage{
translateFils(dimensions_df)
}
\arguments{
\item{filters_df}{a dimensions data frame.}
}
\value{
List of dimensions for the analytics call to datimutils::getAnalytics
}
\description{
takes a filters data frame and translates it into a list item for input to get_analytics
}
