% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{ValidateCodeIdPairs}
\alias{ValidateCodeIdPairs}
\title{ValidateCodeIdPairs}
\usage{
ValidateCodeIdPairs(
  codes,
  ids,
  type,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{codes}{string vector - code of spcific class of metadata - category option, indicator etc}

\item{ids}{string vector - ids of specific class of metadata - category option, indicator etc}

\item{type}{string - metadata endpoint - cataegoryOptions, indicators, etc}
}
\value{
dplyr::all_equal response
}
\description{
Checks code list and paired id list (same length) and verifies they correspond to each other
}
