% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility-Functions.R
\name{diffDataEntryForm}
\alias{diffDataEntryForm}
\title{diffDataEntryForm(uid_a, uid_b)}
\usage{
diffDataEntryForm(
  dataset_a,
  dataset_b,
  sql_view_data_a = NULL,
  sql_view_data_b = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{dataset_a}{a dataset uid}

\item{dataset_b}{a dataset uid}
}
\value{
a list of all three differences
}
\description{
A function that compares two data entry forms based of uid inputs
that access sqlviews
}
