% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{dim_item_sets}
\alias{dim_item_sets}
\title{Input to model calculation gives the mapping between dimension items and category options.}
\format{A data frame with these variables:
\describe{
  \item{dim_uid}{Dimension uid used when running indicators}
  \item{dim_name}{Name for dimension - for documentation not used directly in model calculations}
  \item{dim_item_uid}{Dimension item used when running indicator}
  \item{dim_cop_type}{The dimension type from data pack perspective. These are categories that 
  can appear in the rows of the data pack. In theory different indicators could use different 
  dimensions to produce these disaggregations (e.g. Cascade age bands or Semi fine age). 
  Currently we use the same dimensions respectively for age, sex, kp.}
  \item{dim_item_name}{Name of the dimension item - for documentation not used directly in 
  model calculations}
  \item{option_name}{Name of the category option the dimension item will map to - for 
  documentation not used directly in model calculations}
  \item{option_uid}{UID of the category option the dimension item will map to}
  \item{sort_order}{Sort order for the category option, not strictly necessary for any 
  processing but can be useful in debugging and config work}
  \item{weight}{Value from 0-1 indicating the percent of the value for the dimension item that 
  gets distributed the category option when the raw values are distributed 
  (column "allocate" = "distribute" in data_required.csv) E.g. Dimension item 40-49 may 
  appear twice once for category optiuon 40-44 and once for 45-49. 
  Each of these rows would have weight .5}
  \item{model_sets}{Semicolon seperated list of keys used for grouping the dimension item to 
  category option mappings into sets for calling the analytics api and calculating the model 
  output.}
}}
\source{
COP19 systems team
}
\usage{
dim_item_sets
}
\description{
- Each row corresponds to an individual category option required in the output, 
   there are multiple entries per dimensions and even dimension item in some cases.
 - Includes weight to apply if raw values are being distributed so even a category 
   option can be repeated (<5 and 1-9 data required for 1-4)
 - Defines a key for the different sets of mappings
}
\keyword{datasets}
