% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteDistMain.R
\name{DistributeToSites}
\alias{DistributeToSites}
\title{DistributeToSites()}
\usage{
DistributeToSites(d, data_element_map = datapackcommons::Map19Tto20T,
  mechanisms_historic_global = datapackcommons::Get19TMechanisms("https://www.datim.org/"),
  dim_item_sets = datapackcommons::dim_item_sets, country_name = NULL,
  base_url = getOption("baseurl"), verbose = FALSE)
}
\description{
Takes data pack export (TargetxPSNUxIM) and several other configuration files and 
distributes targets to Site x DSD/TA. Historic data is pulled from the DATIM api and is used to disaggregate 
the PSNUxIM level targets. Maping between New targets and historic data in the data_element_map.
Data element map references dim_item_sets to know which dissags to use and maps from dimension items to category 
options.  
If country_name is passed explicitly and d = NULL then this function returns the distribution 
densities
}
