% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{GetCountryLevels}
\alias{GetCountryLevels}
\title{GetCountryLevels(country_names)}
\usage{
GetCountryLevels(
  countries_req = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{countries_req}{list of country names to include in response}
}
\value{
dataframe with columns country_level, prioritization_level, country_name, id
}
\description{
Gets country uid and prioritization level using dataSetAssignments/ous
}
