% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{GetSqlView}
\alias{GetSqlView}
\title{GetSqlView}
\usage{
GetSqlView(
  sql_view_uid,
  variable_keys = NULL,
  variable_values = NULL,
  col_types = readr::cols(.default = "c"),
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{sql_view_uid}{chr - the uid of the sql view}

\item{variable_keys}{character list - list of the variable names for the sql view}

\item{variable_values}{character list - list of the variable values ordered to correspond with
the related variable key}

\item{col_types}{passed to readr::read_csv col_types parameter for parsing the result set}
}
\value{
dataframe with the results of the sql view
}
\description{
Runs specified sql view with specified sql view variables and returns
a tibble with the results. It is possible to specify the col_types when reading in the data
}
