% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{ValidateCodeIdPairs}
\alias{ValidateCodeIdPairs}
\title{ValidateCodeIdPairs(base_url, codes, ids, type)}
\usage{
ValidateCodeIdPairs(base_url, codes, ids, type)
}
\arguments{
\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{codes}{string vector - code of spcific class of metadata - category option, indicator etc}

\item{ids}{string vector - ids of specific class of metadata - category option, indicator etc}

\item{type}{string - metadata endpoint - cataegoryOptions, indicators, etc}
}
\value{
dplyr::all_equal response
}
\description{
Checks code list and paired id list (same length) and verifies they correspond to each other
}
