% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2api.R
\name{ValidateNameIdPairs}
\alias{ValidateNameIdPairs}
\title{ValidateNameIdPairs(names, ids, type)}
\usage{
ValidateNameIdPairs(base_url, names, ids, type)
}
\arguments{
\item{base_url}{string - base address of instance (text before api/ in URL)}

\item{names}{string vector - names of spcific class of metadata - category option, indicator etc}

\item{ids}{string vector - ids of specific class of metadata - category option, indicator etc}

\item{type}{string - metadata endpoint - cataegoryOptions, indicators, etc}
}
\value{
dplyr::all_equal response
}
\description{
Checks name list and paired id list (same length) and verifies they correspond to each other
}
